moved statuses to utils.ts
